Started initial refactor
