+++ src/Exceptions/PlayerNotFoundException.php
@@ -1,10 +1,4 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: thor
- * Date: 24.06.16
- * Time: 11:19.
- */
 
 namespace TruckersMP\Exceptions;
