+++ src/Helpers/Iteratable.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TruckersMP\Helpers;
+
+
+class Iteratable
+{
+
+}
