+++ src/Models/Servers.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace TruckersMP\Types;
+namespace TruckersMP\Models;
 
-use Psr\Http\Message\ResponseInterface;
 use TruckersMP\Exceptions\APIErrorException;
 
 class Servers implements \Iterator, \ArrayAccess
